+++ src/main/java/com/organization/ordermanagementsystem/helper/consts/CustomerConstants.java
@@ -2,5 +2,5 @@ package com.organization.ordermanagementsystem.helper.consts;
 
 public class CustomerConstants {
     public static final Boolean DEFAULT_DELETED = false;
-    public static final Integer DEFAULT_CUSTOMER_TYPE = 1;
+    public static final Long DEFAULT_CUSTOMER_TYPE = 1L;
 }
