+++ src/main/java/com/organization/ordermanagementsystem/helper/consts/OrdersConstants.java
@@ -4,8 +4,8 @@ import java.math.BigDecimal;
 
 public class OrdersConstants {
     public static final Boolean DEFAULT_DELETED = false;
-    public static final Integer GOLD_CUSTOMER_TYPE = 2;
-    public static final Integer PLATINUM_CUSTOMER_TYPE = 3;
+    public static final Long GOLD_CUSTOMER_TYPE = 2L;
+    public static final Long PLATINUM_CUSTOMER_TYPE = 3L;
     public static final Integer DELIVERY_CHARGE = 50;
     public static final Integer DEFAULT_DISCOUNT_AMOUNT = 0;
 }
